Changed file structure and created namespaces.
