+++ Assets/Game/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44d2de2a8aca4ef6796053eb160133d7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
