+++ Assets/Game/Weapons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96a63baa3a84adec394e023dd009a057
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
