+++ Assets/Scenes/Sandbox.unity
@@ -1473,6 +1473,63 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: c2849fd453abd0445a702a035f04ac60, type: 3}
+--- !u!1001 &2368562753934216616
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2452016873949496323, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_Name
+      value: Enemy
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -10.88
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.46
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 5.23
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6287279941647539006, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: cef93e1fa696b4f08919cb420a011198, type: 3}
 --- !u!1001 &3971697687119871348
 PrefabInstance:
   m_ObjectHideFlags: 0
