+++ Assets/Scripts/Control.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b272adc455912e48a4eda55ad7637d0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
