+++ Assets/Scripts/Core/FollowCamera.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace RPG.Core
+{
+    public class FollowCamera : MonoBehaviour
+    {
+        [SerializeField] Transform target;
+        void LateUpdate()
+        {
+            transform.position = target.position;
+        }
+    }
+}
