--- Assets/Scripts/FollowCamera.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FollowCamera : MonoBehaviour
-{
-    [SerializeField] Transform target;
-    void LateUpdate()
-    {
-        transform.position = target.position;
-    }
-}
