+++ Assets/Scripts/Movement/Mover.cs
@@ -1,0 +1,27 @@
+using UnityEngine;
+using UnityEngine.AI;
+
+namespace RPG.Movement
+{
+    public class Mover : MonoBehaviour
+    {
+        void Update()
+        {
+            updateAnimator();
+        }
+
+        public void MoveTo(Vector3 destination)
+        {
+            GetComponent<NavMeshAgent>().destination = destination;
+        }
+
+        private void updateAnimator()
+        {
+            Vector3 velocity = GetComponent<NavMeshAgent>().velocity;
+            Vector3 localVelocity = transform.InverseTransformDirection(velocity);
+            float speed = localVelocity.z;
+            GetComponent<Animator>().SetFloat("forwardSpeed", speed);
+        }
+    }
+
+}
