+++ Assets/UI/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dac0b85008f0588c6a66d77c154284d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
