+++ Assets/UI/Pause Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5196017aa5d60e81f8b53123ddeffcfd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
